Create matching criteria

Create matching criteria

createMatchingCriteria(controlsPerCase = 1, matchOnAge = TRUE,
  ageCaliper = 2, matchOnGender = TRUE, matchOnProvider = FALSE,
  matchOnCareSite = FALSE, matchOnVisitDate = FALSE,
  visitDateCaliper = 30, matchOnTimeInCohort = FALSE,
  daysInCohortCaliper = 30)

Arguments

controlsPerCase
Maximum number of controls to select per case.

matchOnAge
Match on age?

ageCaliper
Maximum difference (in years) in age when matching on age.

matchOnGender
Match on gender?

matchOnProvider
Match on provider (as specified in the person table)?

matchOnCareSite
Match on care site (as specified in the person table)?

matchOnVisitDate
Should the index date of the control be changed to the nearest visit date?

visitDateCaliper
Maximum difference (in days) between the index date and the visit date when matching on visit date.

matchOnTimeInCohort
Match on time in nesting cohort? When not using nesting, this is interpreted as time observed prior to index.

daysInCohortCaliper
Maximum difference (in days) in time in cohort.